Branch:UserAuth......slight restrucutre of UserRegistrator class, and added basic functionality to add_account() function.
added a username and password attribute and argument for class and constructor
fixed an error in encrypt() function where the base64 encoder was adding a newline to the end of the encryption.
add_accounts() now adds the encrypted username and password to the respective outfile.
need to find the best way to write the initalization vector to the outfile next.